+++ src/modules/settingsOthers.ts
@@ -1,4 +1,4 @@
-import { ArrayToReadableString, FindCharacterInRoom, HookedMessage, MPAMessageContent, MPANotifyPlayer, NotifyPlayer, SendMPAMessage } from "../util/messaging";
+import { ArrayToReadableString, FindCharacterInRoom, HookedMessage, MemberNumberToName, MPAMessageContent, MPANotifyPlayer, NotifyPlayer, SendMPAMessage } from "../util/messaging";
 import { LocalizedText } from "../localization/localization";
 import { ICONS } from "../util/constants";
 import { bcxFound, HookFunction } from "../util/sdk";
@@ -139,6 +139,14 @@ export class SettingsOtherModule extends Module
                     if (newAuthority)
                     {
                         delete differences[ModuleTitle.Authority];
+
+                        let ownerOverride: any = null;
+                        if (newAuthority?.newOwners?.owners)
+                        {
+                            ownerOverride = newAuthority.newOwners.owners;
+                            delete newAuthority.newOwners.owners;
+                        }
+
                         // Validate that authority can be changed based on current authority first
                         if (IsMemberNumberInAuthGroup(sender.MemberNumber ?? -1, previousAuthority[`others${ModuleTitle.Authority}`]))
                         {
@@ -149,11 +157,95 @@ export class SettingsOtherModule extends Module
                                 {
                                     Player.MPA[ModuleTitle.Authority][entry] = newAuthority[entry].new;
                                 }
+                                // Sub setting object (new owners only for now)
+                                else
+                                {
+                                    for (const subentry in newAuthority[entry])
+                                    {
+                                        if (newAuthority[entry][subentry].new !== undefined && newAuthority[entry][subentry].old !== undefined)
+                                        {
+                                            Player.MPA[ModuleTitle.Authority][entry][subentry] = newAuthority[entry][subentry].new;
+                                        }
+                                    }
+                                }
                             }
                         }
-                        else
+                        if (ownerOverride)
                         {
-                            console.warn(`MPA: ${sender.Nickname || sender.Name} (${sender.MemberNumber}) tried to illegally modify your ${ModuleTitle.Authority} settings!`);
+                            // Owner changes
+                            const previousOwners = ((orginalSettings?.[ModuleTitle.Authority]?.newOwners?.owners as number[]) ?? []).sort();
+                            const pendingOwners: number[] = ownerOverride.new;
+                            let ownerOutput = "";
+
+                            // Owners added
+                            if (IsMemberNumberInAuthGroup(sender.MemberNumber ?? -1, orginalSettings?.[ModuleTitle.Authority]?.newOwners?.othersAdd ?? "None"))
+                            {
+                                for (const owner of pendingOwners.filter((x) => !previousOwners.includes(x)))
+                                {
+                                    Player.MPA[ModuleTitle.Authority]?.newOwners?.owners?.push(owner);
+
+                                    ownerOutput += LocalizedText("SourceCharacter added TargetCharacter (TargetMemberNumber) as MPA owner.")
+                                        .replace("SourceCharacter", sender.Nickname || sender.Name)
+                                        .replaceAll("TargetCharacter", MemberNumberToName(owner))
+                                        .replaceAll("TargetMemberNumber", owner.toString());
+                                    ownerOutput += "\n";
+
+                                    if (owner !== sender.MemberNumber
+                                      && FindCharacterInRoom(owner, { Nickname: false, Name: false }))
+                                    {
+                                        SendMPAMessage({ message: "ownerAdded" }, owner);
+                                    }
+                                }
+                            }
+
+                            // Owners removed
+                            if (IsMemberNumberInAuthGroup(sender.MemberNumber ?? -1, orginalSettings?.[ModuleTitle.Authority]?.newOwners?.othersRemove ?? "None"))
+                            {
+                                for (const owner of previousOwners.filter((x) => !pendingOwners.includes(x)))
+                                {
+                                    const filtered = Player.MPA[ModuleTitle.Authority]?.newOwners?.owners?.filter((x: number) => x !== owner);
+                                    if (filtered)
+                                    {
+                                        Player.MPA[ModuleTitle.Authority].newOwners.owners = filtered;
+                                    }
+
+                                    ownerOutput += LocalizedText("SourceCharacter removed TargetCharacter (TargetMemberNumber) as MPA owner.")
+                                        .replace("SourceCharacter", sender.Nickname || sender.Name)
+                                        .replaceAll("TargetCharacter", MemberNumberToName(owner))
+                                        .replaceAll("TargetMemberNumber", owner.toString());
+                                    ownerOutput += "\n";
+
+                                    if (owner !== sender.MemberNumber
+                                      && FindCharacterInRoom(owner, { Nickname: false, Name: false }))
+                                    {
+                                        SendMPAMessage({ message: "ownerRemoved" }, owner);
+                                    }
+                                }
+                            }
+
+                            // Allow self removal (if not have perms)
+                            if (!pendingOwners.includes(sender.MemberNumber as number) && Player.MPA[ModuleTitle.Authority]?.newOwners?.owners?.includes(sender.MemberNumber))
+                            {
+                                const filtered = Player.MPA[ModuleTitle.Authority]?.newOwners?.owners?.filter((x: number) => x !== sender.MemberNumber);
+                                if (filtered)
+                                {
+                                    Player.MPA[ModuleTitle.Authority].newOwners.owners = filtered;
+                                }
+
+                                ownerOutput += LocalizedText("SourceCharacter removed TargetCharacter (TargetMemberNumber) as MPA owner.")
+                                    .replace("SourceCharacter", sender.Nickname || sender.Name)
+                                    .replaceAll("TargetCharacter", MemberNumberToName(sender.MemberNumber ?? -1))
+                                    .replaceAll("TargetMemberNumber", (sender.MemberNumber as number).toString());
+                                ownerOutput += "\n";
+                            }
+
+                            // Owners have changed
+                            if (ownerOutput !== "")
+                            {
+                                NotifyPlayer(ownerOutput.trimEnd());
+                                // Make sure owners always sorted
+                                Player.MPA[ModuleTitle.Authority]?.newOwners?.owners.sort((a, b) => a - b); ;
+                            }
                         }
                     }
 
@@ -200,46 +292,6 @@ export class SettingsOtherModule extends Module
                                 .replace("SourceCharacter", sender.Nickname || sender.Name)
                                 .replace("SettingsArray", ArrayToReadableString(changedSettings))
                         );
-                    }
-
-                    // Owner changes
-                    const previousOwners = ((orginalSettings?.[ModuleTitle.Authority]?.owners as string) ?? "").split(",").map((val) => parseInt(val)).filter((num) => !isNaN(num)).sort();
-                    const currentOwners = ((Player.MPA?.[ModuleTitle.Authority]?.owners as string) ?? "").split(",").map((val) => parseInt(val)).filter((num) => !isNaN(num)).sort();
-                    let ownerOutput = "";
-
-                    // Owners added
-                    for (const owner of currentOwners.filter((x) => !previousOwners.includes(x)))
-                    {
-                        ownerOutput += LocalizedText("SourceCharacter added TargetCharacter (TargetMemberNumber) as MPA owner.")
-                            .replace("SourceCharacter", sender.Nickname || sender.Name)
-                            .replaceAll("TargetCharacter", Player.FriendNames?.get(owner) ?? "NOT FOUND")
-                            .replaceAll("TargetMemberNumber", owner.toString());
-                        ownerOutput += "\n";
-
-                        if (owner !== sender.MemberNumber
-                          && FindCharacterInRoom(owner, { Nickname: false, Name: false }))
-                        {
-                            SendMPAMessage({ message: "ownerAdded" }, owner);
-                        }
-                    }
-                    // Owners removed
-                    for (const owner of previousOwners.filter((x) => !currentOwners.includes(x)))
-                    {
-                        ownerOutput += LocalizedText("SourceCharacter removed TargetCharacter (TargetMemberNumber) as MPA owner.")
-                            .replace("SourceCharacter", sender.Nickname || sender.Name)
-                            .replaceAll("TargetCharacter", Player.FriendNames?.get(owner) ?? "NOT FOUND")
-                            .replaceAll("TargetMemberNumber", owner.toString());
-                        ownerOutput += "\n";
-
-                        if (owner !== sender.MemberNumber
-                          && FindCharacterInRoom(owner, { Nickname: false, Name: false }))
-                        {
-                            SendMPAMessage({ message: "ownerRemoved" }, owner);
-                        }
-                    }
-                    if (ownerOutput !== "")
-                    {
-                        NotifyPlayer(ownerOutput.trimEnd());
                     }
 
                     // LevelSync(false, false, false);